refactor: isNullSquareExistをリフトアップ
