+++ src/model.ts
@@ -53,3 +53,16 @@ export function calculateWinner(board: string[][]): number[][] | null {
   }
   return null;
 }
+
+export function checkNullSquareExist(Board: string[][]): boolean {
+  let isNullSquareExist = false;
+  Board.forEach((nthRow) => {
+    nthRow.forEach((nthSquare) => {
+      if (!nthSquare) {
+        isNullSquareExist = true;
+      }
+    });
+  });
+
+  return isNullSquareExist;
+}
